rename settings/navigation -> settings/design, move theme management
no issue
- following from the `settings/general` redesign, navigation and themes management are now co-located in `settings/design`
