+++ app/controllers/settings/design.js
@@ -1,15 +1,22 @@
 import RSVP from 'rsvp';
 import Controller from 'ember-controller';
-import computed from 'ember-computed';
+import computed, {notEmpty} from 'ember-computed';
 import injectService from 'ember-service/inject';
 import SettingsSaveMixin from 'ghost-admin/mixins/settings-save';
 import NavigationItem from 'ghost-admin/models/navigation-item';
+import $ from 'jquery';
 
 export default Controller.extend(SettingsSaveMixin, {
     config: injectService(),
+    ghostPaths: injectService(),
     notifications: injectService(),
+    session: injectService(),
 
     newNavItem: null,
+
+    themes: null,
+    themeToDelete: null,
+    showDeleteThemeModal: notEmpty('themeToDelete'),
 
     blogUrl: computed('config.blogUrl', function () {
         let url = this.get('config.blogUrl');
@@ -29,7 +36,7 @@ export default Controller.extend(SettingsSaveMixin, {
         let validationPromises = [];
 
         if (!newNavItem.get('isBlank')) {
-            validationPromises.pushObject(this.send('addItem'));
+            validationPromises.pushObject(this.send('addNavItem'));
         }
 
         navItems.map((item) => {
@@ -54,8 +61,20 @@ export default Controller.extend(SettingsSaveMixin, {
         this.set('newNavItem', NavigationItem.create({isNew: true}));
     },
 
+    _deleteTheme() {
+        let theme = this.get('store').peekRecord('theme', this.get('themeToDelete').name);
+
+        if (!theme) {
+            return;
+        }
+
+        return theme.destroyRecord().catch((error) => {
+            this.get('notifications').showAPIError(error);
+        });
+    },
+
     actions: {
-        addItem() {
+        addNavItem() {
             let newNavItem = this.get('newNavItem');
 
             // If the url sent through is blank (user never edited the url)
@@ -68,7 +87,7 @@ export default Controller.extend(SettingsSaveMixin, {
             });
         },
 
-        deleteItem(item) {
+        deleteNavItem(item) {
             if (!item) {
                 return;
             }
@@ -88,6 +107,36 @@ export default Controller.extend(SettingsSaveMixin, {
             }
 
             navItem.set('url', url);
+        },
+
+        setTheme(theme) {
+            this.set('model.activeTheme', theme.name);
+            this.send('save');
+        },
+
+        downloadTheme(theme) {
+            let themeURL = `${this.get('ghostPaths.apiRoot')}/themes/${theme.name}`;
+            let accessToken = this.get('session.data.authenticated.access_token');
+            let downloadURL = `${themeURL}/download/?access_token=${accessToken}`;
+            let iframe = $('#iframeDownload');
+
+            if (iframe.length === 0) {
+                iframe = $('<iframe>', {id: 'iframeDownload'}).hide().appendTo('body');
+            }
+
+            iframe.attr('src', downloadURL);
+        },
+
+        deleteTheme(theme) {
+            if (theme) {
+                return this.set('themeToDelete', theme);
+            }
+
+            return this._deleteTheme();
+        },
+
+        hideDeleteThemeModal() {
+            this.set('themeToDelete', null);
         },
 
         reset() {
