+++ app/router.js
@@ -42,16 +42,16 @@ GhostRouter.map(function () {
         this.route('user', {path: ':user_slug'});
     });
 
-    this.route('settings.general', {path: '/settings/general'}, function () {
-        this.route('uploadtheme');
-    });
+    this.route('settings.general', {path: '/settings/general'});
     this.route('settings.tags', {path: '/settings/tags'}, function () {
         this.route('tag', {path: ':tag_slug'});
         this.route('new');
     });
     this.route('settings.labs', {path: '/settings/labs'});
     this.route('settings.code-injection', {path: '/settings/code-injection'});
-    this.route('settings.navigation', {path: '/settings/navigation'});
+    this.route('settings.design', {path: '/settings/design'}, function () {
+        this.route('uploadtheme');
+    });
     this.route('settings.apps', {path: '/settings/apps'}, function () {
         this.route('slack', {path: 'slack'});
         this.route('amp', {path: 'amp'});
