+++ app/templates/components/gh-nav-menu.hbs
@@ -51,7 +51,7 @@
             <li class="gh-nav-list-h">Settings</li>
             <li>{{#link-to "settings.general" classNames="gh-nav-settings-general"}}<i class="icon-settings"></i>General{{/link-to}}</li>
             {{!<li><i class="icon-design"></i>Themes</li>}}
-            <li>{{#link-to "settings.navigation" classNames="gh-nav-settings-navigation"}}<i class="icon-compass"></i>Navigation{{/link-to}}</li>
+            <li>{{#link-to "settings.design" classNames="gh-nav-settings-navigation"}}<i class="icon-compass"></i>Design{{/link-to}}</li>
             <li>{{#link-to "settings.tags" classNames="gh-nav-settings-tags"}}<i class="icon-tag"></i>Tags{{/link-to}}</li>
             <li>{{#link-to "settings.code-injection" classNames="gh-nav-settings-code-injection"}}<i class="icon-code"></i>Code injection{{/link-to}}</li>
             <li>{{#link-to "settings.apps" classNames="gh-nav-settings-apps"}}<i class="icon-box"></i>Apps{{/link-to}}</li>
