+++ app/templates/settings/design-loading.hbs
@@ -1,6 +1,6 @@
 <section class="gh-canvas">
     <header class="gh-canvas-header">
-        <h2 class="gh-canvas-title">Navigation</h2>
+        <h2 class="gh-canvas-title">Design</h2>
         <section class="view-actions">
             {{#gh-spin-button class="gh-btn gh-btn-blue" action="save" submitting=submitting}}<span>Save</span>{{/gh-spin-button}}
         </section>
