+++ app/templates/settings/design/uploadtheme.hbs
