--- app/templates/settings/navigation.hbs
@@ -1,19 +1,0 @@
-<section class="gh-canvas">
-    <header class="gh-canvas-header">
-        <h2 class="gh-canvas-title">Navigation</h2>
-        <section class="view-actions">
-            {{#gh-spin-button class="gh-btn gh-btn-blue" action="save" submitting=submitting}}<span>Save</span>{{/gh-spin-button}}
-        </section>
-    </header>
-
-    <section class="view-container">
-        <form id="settings-navigation" class="gh-blognav" novalidate="novalidate">
-            {{#sortable-group onChange=(action 'reorderItems') as |group|}}
-                {{#each model.navigation as |navItem|}}
-                    {{gh-navitem navItem=navItem baseUrl=blogUrl addItem="addItem" deleteItem="deleteItem" updateUrl="updateUrl" group=group}}
-                {{/each}}
-            {{/sortable-group}}
-            {{gh-navitem navItem=newNavItem baseUrl=blogUrl addItem="addItem" updateUrl="updateUrl"}}
-        </form>
-    </section>
-</section>
